the signin – loginPage is changed to the index of the app and few files in modules were also modified.
The header and carousel components were also modified.

The model module was renamed to listeningHistory module for users listening History table on the database.

Setup states for the AudioPlay component and users listening history and context for the playerRef object that is being destructured from the context.

created a AudioPlay component taking audioFile and audioRef props and audio HTML element used to play audio files is returned with ref, src, controls and autoPlay attributes.