creating a state variable history with an initial state of null. setHistory is the function to update this state.
creating a reference playerRef to an HTML audio element. This reference can be attached to an audio element in the JSX, and it allows you to interact with that element programmatically.

In the useEffect hook used to perform side effects and fetch the users favorites from the database when the component mounts, the history data fetch was also added to fetch user_history table.