Created a ListeningHistoryPage component to mapping over the history state from context and rendering users Listening History details for each item in the history.
In the show details page, state variables are created for playStatus, currentFile and playProgress with own initial states and their associated functions to update states.

Created a function handlePlayEpisode that handles the logic for playing an episode. This is handled by the Listen button.

Added an AppBar component at the bottom of the show details page with a Toolbar component with a marquee that displays the title of the current file and an AudioPlayer component that plays the current episode file.