created favoriteEpisodes state so users favorited episodes ca be remembered across different sessions of the App.
stored the favoriteEpisodes state in the global context (ShowsContext).

Stored IDs of the favorited episodes in a state variable which updates whenever the user adds or removes an episode from their favorites.

Added a new state variable to keep track of the favorited episodes the included favoriteEpisodes and setFavoriteEpisodes in the value provided by the ShowsContext.Provider.

Used setFavoriteEpisodes function from the context to remove an episode from the favorites in the Favourite component.

created handleRemoveFavorite function to handle the removal of an episode from the favorites and updated the remove button of the Favourite component.